Split redux_validation file
Now each class has its own file.
