+++ app/js/lib/validation_dispatcher.js
@@ -42,19 +42,6 @@ var objectAssign = require( 'object-assign' ),
 		}
 	},
 
-	ValidationDispatcherCollection = {
-		dispatchers: [],
-		store: null,
-		formContentName: '',
-		onUpdate: function () {
-			var formContent = this.store.getState()[ this.formContentName ],
-				i;
-			for ( i = 0; i < this.dispatchers.length; i++ ) {
-				this.dispatchers[ i ].dispatchIfChanged( formContent, this.store );
-			}
-		}
-	},
-
 	/**
 	 *
 	 * @param {Function|Object} validator Function or object that has a 'validate' method.
@@ -74,25 +61,8 @@ var objectAssign = require( 'object-assign' ),
 			fields: fieldNames,
 			previousFieldValues: _.pick( initialValues || {}, fieldNames )
 		} );
-	},
-
-	/**
-	 *
-	 * @param {Object} store Redux store
-	 * @param {ValidationDispatcher[]} dispatchers
-	 * @param {string} formContentName Field name for the store to access form contents, e.g. 'donationFormContent' or 'membershipFormContent'
-	 */
-	createValidationDispatcherCollection = function ( store, dispatchers, formContentName ) {
-		var collection = objectAssign( Object.create( ValidationDispatcherCollection ), {
-				store: store,
-				dispatchers: dispatchers,
-				formContentName: formContentName
-			} );
-		store.subscribe( collection.onUpdate.bind( collection ) );
-		return collection;
 	};
 
 module.exports = {
-	createValidationDispatcher: createValidationDispatcher,
-	createValidationDispatcherCollection: createValidationDispatcherCollection
+	createValidationDispatcher: createValidationDispatcher
 };
